refactor: 스타일로 분리
