+++ src/Router.tsx
@@ -1,11 +1,11 @@
 import { createBrowserRouter } from "react-router-dom";
 import Layout from "./pages/Layout";
 import Home from "./pages/home/Home";
-import Scrap from "./pages/scrap/Scrap";
+
 import Complaint from "./pages/my-complaint/Complaint";
 import Detail from "./pages/detail/Detail";
 import View from "./pages/view/View";
-import AdminHome from "./pages/AdminHome"
+import AdminHome from "./pages/AdminHome";
 import MyPage from "./pages/my-page/MyPage";
 import Search from "./pages/search/Search";
 
@@ -19,16 +19,12 @@ export const router = createBrowserRouter([
         element: <Home />,
       },
       {
-        path: "my-scrap",
-        element: <Scrap />,
-      },
-      {
         path: "mypage/my-complaint",
         element: <Complaint />,
       },
+      {
         path: "complaint-detail",
         element: <Detail />,
-        },
       },
       {
         path: "complaint-search",
@@ -45,7 +41,7 @@ export const router = createBrowserRouter([
     ],
   },
   {
-    path: "admin",
+    path: "/admin",
     children: [
       {
         path: "home",
