+++ src/pages/my-complaint/Complaint.tsx
@@ -1,69 +1,28 @@
-import styled from "@emotion/styled";
 import ContentBox from "../../components/content/ContentBox";
 import { mockData } from "../../mockData";
 import { motion } from "framer-motion";
-import SvgIcon, { SvgIconProps } from "@mui/material/SvgIcon";
 import ContactPageRoundedIcon from "@mui/icons-material/ContactPageRounded";
 import FilterBar from "./ComplaintFilterBar";
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  gap: 3rem;
-`;
-
-const TitleContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
-
-const ComplaintIcon = styled(SvgIcon)<SvgIconProps>`
-  color: var(--disabled-primary);
-  width: 2rem;
-  height: 2rem;
-  margin-top: 2rem;
-`;
-
-const ComplaintTitle = styled.h2`
-  color: var(--gray6-black);
-`;
-
-const ComplaintBorder = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-top: 1px solid var(--gray3-border);
-  padding: 5rem;
-  transform: translateY(-5rem);
-  z-index: 0;
-  width: 100vw;
-  flex-wrap: wrap;
-`;
-
-//complaint를 시작점부터 배치
-const ComplaintGrid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 2rem;
-  justify-content: start;
-  align-items: center;
-`;
+import {
+  Container,
+  TitleContainer,
+  ComplaintIcon,
+  Title,
+  Border,
+  ComplaintGrid,
+} from "../../styles/ComplaintScrap";
 
 const Complaint = () => {
   return (
     <Container>
       <TitleContainer>
         <ComplaintIcon component={ContactPageRoundedIcon} />
-        <ComplaintTitle>내 민원</ComplaintTitle>
+        <Title>내 민원</Title>
       </TitleContainer>
 
       <FilterBar />
 
-      <ComplaintBorder>
+      <Border>
         <ComplaintGrid>
           {mockData.map((item, index) => (
             <motion.div
@@ -76,7 +35,7 @@ const Complaint = () => {
             </motion.div>
           ))}
         </ComplaintGrid>
-      </ComplaintBorder>
+      </Border>
     </Container>
   );
 };
